+++ tests/sdk/split_log_test.go
@@ -1,0 +1,103 @@
+package sdk
+
+import (
+	"github.com/starkbank/sdk-go/starkbank"
+	SplitLog "github.com/starkbank/sdk-go/starkbank/split/log"
+	Utils "github.com/starkbank/sdk-go/tests/utils"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestSplitLogGet(t *testing.T) {
+
+	starkbank.User = Utils.ExampleProject
+
+	limit := 10
+	var params = map[string]interface{}{}
+	params["limit"] = limit
+
+	var logIds, splitIds []string
+
+	logs, errorChannel := SplitLog.Query(params, nil)
+	loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			logIds = append(logIds, log.Id)
+		}
+	}
+
+	for _, ids := range logIds {
+		splitLog, err := SplitLog.Get(ids, nil)
+		if err.Errors != nil {
+			for _, erro := range err.Errors {
+				t.Errorf("code: %s, message: %s", erro.Code, erro.Message)
+			}
+		}
+		assert.NotNil(t, splitLog)
+		splitIds = append(splitIds, splitLog.Id)
+	}
+	assert.Equal(t, logIds, splitIds)
+}
+
+func TestSplitLogQuery(t *testing.T) {
+
+	starkbank.User = Utils.ExampleProject
+
+	limit := 10
+	var params = map[string]interface{}{}
+	params["limit"] = limit
+
+	var logList []SplitLog.Log
+
+	logs, errorChannel := SplitLog.Query(params, nil)
+	loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			logList = append(logList, log)
+		}
+	}
+
+	assert.Equal(t, limit, len(logList))
+}
+
+func TestSplitLogPage(t *testing.T) {
+
+	starkbank.User = Utils.ExampleProject
+
+	var ids []string
+	var params = map[string]interface{}{}
+	params["limit"] = 4
+
+	logs, cursor, err := SplitLog.Page(params, nil)
+	for _, log := range logs {
+		if err.Errors != nil {
+			for _, erro := range err.Errors {
+				t.Errorf("code: %s, message: %s", erro.Code, erro.Message)
+			}
+		}
+		ids = append(ids, log.Id)
+		assert.NotNil(t, log.Id)
+		assert.NotNil(t, cursor)
+	}
+	assert.Len(t, ids, 4)
+}